Collision bug (#16)
* copied airfoil into new directory fabric, and added to the build; will be used to make libFabric.la with external fluid solver

* removed iFluid.h from all include statements, added ifluid_state.h to directory and included in airfoil.h temporarily

* need to figure out what components of the propagation functions stay and what goes

* removed 2d functions

* added state.h (formerly ifluid_state.h)

* added F_PARAMS and F_FIELD to airfoil.h

* create fluidprop.cpp and reorganized functions

* transferred necessary fluid functions into fluidprop.cpp -- a temporary measure

* stripped iFluid solver out of driver

* copied in collision files, and set Makefile.am to compile libfabric.la

* August (#13)

* brackets and indent for clarity in if statement

* complex connection needs to be set to no for in-DGB0 to work

* removed openmp pragma

* added input-DGB directory for parachute

* updated input-C9 file defaults

* updated config_omega to match config_alpha after OS upgrade

* qrun-inputdir* scripts use nodes.q group instead of all.q, prevents running non-gpu jobs on the gpu node

* almost have libfabric.la completely contained in the fabric directory, aside from libFronTier.la

* rename program to fabric.cpp

* added rigidbody.h and rigidbody.cpp

* rigidbody routines nearly finished

* got fabric library to compile seperate from all other modules; still need to decouple from fluid data structures within the fabric directory

* removed commented out function prototype

* August (#14)

* brackets and indent for clarity in if statement

* complex connection needs to be set to no for in-DGB0 to work

* removed openmp pragma

* added input-DGB directory for parachute

* updated input-C9 file defaults

* updated config_omega to match config_alpha after OS upgrade

* qrun-inputdir* scripts use nodes.q group instead of all.q, prevents running non-gpu jobs on the gpu node

* removed DAMPING macros

* removed DAMPING macros from libFabric.la files

* added X.Li changes

* added test1 executable

* removed input-2d dir

* added in-drum* files for testsing

* added X.Li fabric_api functions and test driver

* added extern C statements to fabric_api.cpp for interoperation with fortran

* removed unneccesary extern keywords

* added fdecs.F90

* making initialization api calls no ops; finished f_basic_data type in fdecs.F90, may need later

* added fabric.F90 test driver

* added fortrantest executable to Makfile.am

* added fortran test driver; still in development

* added missing //c_null_char to in_name char array in fortran_driver.F90 -- bug fixed

* removed dead/commented code

* still not working

* added frun shell script as a tempory measure so development can continue -- will come up with better solution for passing input files

* simple (temportary) fortran driver now working

* saving old directory temporarily as reference

* added X.Li new code

* merge X.Li code for test.cpp and made corresponding fortran driver

* removed fabric-save after the merge

* added FT_InteriorPropagate() calls to parachute.cpp driver

* removed commented out function

* added FT_InteriorPropagate() calls to airfoil driver to match those in the parachute driver

* update fgb3dutil.c X.Li

* added damping terms in afset.cpp

* added damping to fabric/afsetd.cpp

* new input-drum directory

* added FT_InteriorPropagate() calls to fabric.cpp

* remove old comment, add comment about redundant code

* commented out reduceSuperelast()

* corrections to input-collid-new input files

* additional comments

* added variable to AF_PARAMS for collision detection -- match tengbo's version of code

* added options for setting some collision parametes by input file -- from tengbo's code

* match tengbo's collision code, biggest change is removing SVD for ill conditioned inertia tensors in dcollid3d.cpp

* added missing declarations to collid.h

* added commented out chard coded variables that was different in tengbo's version, not sure what it is for -- idiagnostic.c

* copied in tengbo's lates AABB code

* clean_up stack trace debugging

* input file to test stacktrace on clean_up

* changed friction to more realilistic 0.01 in input file

* commented out backtrace_symbols_fd I'm using to debug the stacktrace in clean_up()

* added some comments regarding artificial viscosity/damping for future implementation

* don't go into computeImpactZone() if there is only a small number of collision pairs (even if over MAX_ITERS)

+++ src/intfc/igrid.c
@@ -1247,7 +1247,9 @@ EXPORT boolean rect_in_which(
 	const int   *gmax = grid->gmax;
 	const int   *lbuf = grid->lbuf, *ubuf = grid->ubuf;
 	int	    i, dim = grid->dim;
-	int SHIFT = 0.2; /* TOLERANCE */
+	
+    double SHIFT = 0.2; /* TOLERANCE */
+        //int SHIFT = 0.2; /* TOLERANCE */
 
 		/* Find Grid Block and points outside and moved in */
